+++ source/editor/utils/ObjectIcons.js
@@ -77,7 +77,6 @@ ObjectIcons.icons = new Map([
 	["Group", ObjectIcons.path + "misc/container.png"],
 	["BillboardGroup", ObjectIcons.path + "misc/rotate.png"],
 	["LensFlare", ObjectIcons.path + "misc/flare.png"],
-	["OrbitControls", ObjectIcons.path + "misc/orbit.png"],
 	["FirstPersonControls", ObjectIcons.path + "misc/crosshair.png"]
 ]);